refactor: 회원가입 페이지의 래핑 컴포넌트를 template으로 분리
