+++ src/pages/Signup/index.tsx
@@ -7,11 +7,7 @@ import { ViewIcon, ViewOffIcon } from '@chakra-ui/icons';
 import useSignupMutation from '@/apis/mutations/useSignupMutation';
 import musseuk from '@/assets/images/musseuk_hood.png';
 import InputField from './components/InputField';
-
-const colors = {
-  background: '#F2F8EB',
-  submit: '#8CD790'
-};
+import PageTemplate from './templates/PageTemplate';
 
 const links = {
   main: '/',
@@ -74,112 +70,95 @@ const SignUp = () => {
   const onError: SubmitErrorHandler<Inputs> = (errors) => console.error(errors);
 
   return (
-    <Flex w="100%" px="4" py="8" minH="100vh" bg={colors.background} justifyContent="center" alignItems="center">
-      <Flex
-        as="form"
-        onSubmit={handleSubmit(onSubmit, onError)}
-        px={['8', '16', '24']}
-        py={['6', '8', '12']}
+    <PageTemplate onSubmit={handleSubmit(onSubmit, onError)}>
+      <Image maxW="32" src={musseuk} alt="머쓱이" />
+      <Heading textAlign="center">Sign up</Heading>
+      <InputField
+        label="Email"
+        error={errors.email}
+        inputProps={{
+          id: 'email',
+          type: 'email',
+          placeholder: '이메일을 입력해주세요'
+        }}
+        registerProps={register('email')}
+      />
+      <InputField
+        label="Username"
+        error={errors.username}
+        inputProps={{
+          id: 'username',
+          type: 'text',
+          placeholder: '실명을 입력해주세요',
+          maxLength: 4
+        }}
+        registerProps={register('username')}
+      />
+      <InputField
+        label="Password"
+        inputProps={{
+          id: 'password',
+          type: showPassword ? 'text' : 'password',
+          placeholder: '비밀번호를 입력해주세요',
+          maxLength: 30
+        }}
+        registerProps={register('password')}
+        icon={<Icon as={showPassword ? ViewOffIcon : ViewIcon} onClick={setShowPassword.toggle} />}
+      />
+      <Box w="100%" fontSize="sm">
+        {isPasswordShort && <Text>· Length must be greater than 8 characters</Text>}
+        {!isPasswordContainNumber && <Text color="green03">· Password must contain numbers</Text>}
+      </Box>
+      <InputField
+        label="Confirm Password"
+        error={errors.confirmPassword}
+        inputProps={{
+          id: 'confirm-password',
+          type: showConfirmPassword ? 'text' : 'password',
+          placeholder: '비밀번호를 재확인해주세요',
+          maxLength: 30
+        }}
+        registerProps={register('confirmPassword')}
+        icon={<Icon as={showConfirmPassword ? ViewOffIcon : ViewIcon} onClick={setShowConfirmPassword.toggle} />}
+      />
+      <Button
+        type="submit"
+        mt="6"
         w="100%"
-        maxW="xl"
-        bg="white"
-        boxSizing="border-box"
-        boxShadow="lg"
-        borderRadius="20px"
-        flexDir="column"
-        gap="2"
-        justifyContent="center"
+        color="white"
+        backgroundColor="green01"
+        _hover={{ backgroundColor: 'green03' }}
+        _active={{ backgroundColor: 'green.500' }}>
+        Create account
+      </Button>
+      <Flex
+        mt="3"
+        w="100%"
+        direction={['column', 'row']}
+        justifyContent="space-around"
         alignItems="center"
-        transition="padding 0.25s ease-in-out">
-        <Image maxW="32" src={musseuk} alt="머쓱이" />
-        <Heading textAlign="center">Sign up</Heading>
-        <InputField
-          label="Email"
-          error={errors.email}
-          inputProps={{
-            id: 'email',
-            type: 'email',
-            placeholder: '이메일을 입력해주세요'
-          }}
-          registerProps={register('email')}
-        />
-        <InputField
-          label="Username"
-          error={errors.username}
-          inputProps={{
-            id: 'username',
-            type: 'text',
-            placeholder: '실명을 입력해주세요',
-            maxLength: 4
-          }}
-          registerProps={register('username')}
-        />
-        <InputField
-          label="Password"
-          inputProps={{
-            id: 'password',
-            type: showPassword ? 'text' : 'password',
-            placeholder: '비밀번호를 입력해주세요',
-            maxLength: 30
-          }}
-          registerProps={register('password')}
-          icon={<Icon as={showPassword ? ViewOffIcon : ViewIcon} onClick={setShowPassword.toggle} />}
-        />
-        <Box w="100%" fontSize="sm">
-          {isPasswordShort && <Text>· Length must be greater than 8 characters</Text>}
-          {!isPasswordContainNumber && <Text color="green03">· Password must contain numbers</Text>}
-        </Box>
-        <InputField
-          label="Confirm Password"
-          error={errors.confirmPassword}
-          inputProps={{
-            id: 'confirm-password',
-            type: showConfirmPassword ? 'text' : 'password',
-            placeholder: '비밀번호를 재확인해주세요',
-            maxLength: 30
-          }}
-          registerProps={register('confirmPassword')}
-          icon={<Icon as={showConfirmPassword ? ViewOffIcon : ViewIcon} onClick={setShowConfirmPassword.toggle} />}
-        />
-        <Button
-          type="submit"
-          mt="6"
-          w="100%"
-          color="white"
-          backgroundColor="green01"
-          _hover={{ backgroundColor: 'green03' }}
-          _active={{ backgroundColor: 'green.500' }}>
-          Create account
-        </Button>
-        <Flex
-          mt="3"
-          w="100%"
-          direction={['column', 'row']}
-          justifyContent="space-around"
-          alignItems="center"
-          transition="color 0.2s"
-          gap="1"
-          fontSize="sm"
-          textAlign="center">
-          <Text color="gray.400">Already have an account?</Text>
-          <Link to={links.signin}>
-            <Text
-              color="green.500"
-              _hover={{
-                color: 'green.600'
-              }}
-              _active={{
-                color: 'green.700'
-              }}
-              fontWeight="semibold"
-              cursor="pointer"
-              userSelect="none">
-              Sign in
-            </Text>
-          </Link>
-        </Flex>
+        transition="color 0.2s"
+        gap="1"
+        fontSize="sm"
+        textAlign="center">
+        <Text color="gray.400">Already have an account?</Text>
+        <Link to={links.signin}>
+          <Text
+            color="green.500"
+            _hover={{
+              color: 'green.600'
+            }}
+            _active={{
+              color: 'green.700'
+            }}
+            fontWeight="semibold"
+            cursor="pointer"
+            userSelect="none">
+            Sign in
+          </Text>
+        </Link>
       </Flex>
-    </Flex>
+    </PageTemplate>
   );
 };
